Smart contract controller removed, added as a simple test utility class.
