+++ src/smartcontracts/index.ts
@@ -8,7 +8,6 @@ export * from "./argSerializer";
 export * from "./code";
 export * from "./codec";
 export * from "./codeMetadata";
-export * from "./smartContractController";
 export * from "./function";
 export * from "./interaction";
 export * from "./interactionChecker";
